first commita

+++ lib/user2/features/Customer/presentation/widgets/totalPaymentDialog.dart
@@ -7,7 +7,7 @@ class PaymentDialog extends StatefulWidget {
   final String customerName;
   final List<Bill> customerBills;
   final BillRepository billRepository;
-  final Function(int, int,int) updateBillStatus;
+  final Function(int, double,double) updateBillStatus;
 
   PaymentDialog({
     required this.customerName,
@@ -24,7 +24,7 @@ class _PaymentDialogState extends State<PaymentDialog> {
   final TextEditingController _totalPaymentController = TextEditingController();
   List<Map<String, String>> _vaults = [];
   String? _selectedVaultId;
-  Map<int, int> _updatedPayments = {};
+  Map<int, double> _updatedPayments = {};
   bool _isDistributed = false;
 
   @override
@@ -36,7 +36,7 @@ class _PaymentDialogState extends State<PaymentDialog> {
 
   void _initializePayments() {
     _updatedPayments = {
-      for (var bill in widget.customerBills) bill.id: bill.payment ?? 0
+      for (var bill in widget.customerBills) bill.id: bill.payment ?? 0.0
     };
   }
 
@@ -50,7 +50,7 @@ class _PaymentDialogState extends State<PaymentDialog> {
     }
   }
 
-  void _applyPayment(int enteredAmount) {
+  void _applyPayment(double enteredAmount) {
     if (enteredAmount <= 0) {
       ScaffoldMessenger.of(context).showSnackBar(
         SnackBar(content: Text('يرجى إدخال مبلغ صالح')),
@@ -58,19 +58,19 @@ class _PaymentDialogState extends State<PaymentDialog> {
       return;
     }
 
-    int remainingPayment = enteredAmount;
+    double remainingPayment = enteredAmount;
 
     for (var bill in widget.customerBills) {
-      if (remainingPayment <= 0) break;
+      if (remainingPayment <= 0.0) break;
 
-      int remainingAmountForBill = bill.total_price - _updatedPayments[bill.id]!;
+      double remainingAmountForBill = bill.total_price - _updatedPayments[bill.id]!;
 
       if (remainingPayment >= remainingAmountForBill) {
         _updatedPayments[bill.id] = bill.total_price;
         remainingPayment -= remainingAmountForBill;
       } else {
         _updatedPayments[bill.id] = _updatedPayments[bill.id]! + remainingPayment;
-        remainingPayment = 0;
+        remainingPayment = 0.0;
       }
     }
 
@@ -82,9 +82,9 @@ class _PaymentDialogState extends State<PaymentDialog> {
     setState(() => _isDistributed = true);
   }
 
-  int _getTotalAmount() => widget.customerBills.fold(0, (sum, bill) => sum + bill.total_price);
-  int _getTotalPaid() => widget.customerBills.fold(0, (sum, bill) => sum + _updatedPayments[bill.id]!);
-  int _getRemainingAmount() => _getTotalAmount() - _getTotalPaid();
+  double _getTotalAmount() => widget.customerBills.fold(0, (sum, bill) => sum + bill.total_price);
+  double _getTotalPaid() => widget.customerBills.fold(0, (sum, bill) => sum + _updatedPayments[bill.id]!);
+  double _getRemainingAmount() => _getTotalAmount() - _getTotalPaid();
 
   @override
   Widget build(BuildContext context) {
@@ -105,7 +105,7 @@ class _PaymentDialogState extends State<PaymentDialog> {
           SizedBox(height: 10),
           ElevatedButton(
             onPressed: () {
-              int enteredAmount = int.tryParse(_totalPaymentController.text) ?? 0;
+              double enteredAmount = double.tryParse(_totalPaymentController.text) ?? 0;
               _applyPayment(enteredAmount);
             },
             child: Text('توزيع المبلغ'),
@@ -142,7 +142,7 @@ class _PaymentDialogState extends State<PaymentDialog> {
 
   Future<void> _savePayments() async {
     for (var bill in widget.customerBills) {
-      int newPayment = _updatedPayments[bill.id]!;
+      double newPayment = _updatedPayments[bill.id]!;
       if (newPayment > bill.payment!) {
         await widget.billRepository.addPayment(
           billId: bill.id,
@@ -160,7 +160,7 @@ class _PaymentDialogState extends State<PaymentDialog> {
     Navigator.pop(context, true);
   }
 
-  Widget _buildSummaryInfo(int total, int paid, int remaining) {
+  Widget _buildSummaryInfo(double total, double paid, double remaining) {
     return Column(
       children: [
         _buildInfoText('إجمالي الفواتير', total,Colors.blue),
@@ -170,14 +170,14 @@ class _PaymentDialogState extends State<PaymentDialog> {
     );
   }
 
-  Widget _buildInfoText(String label, int amount, Color? color) {
+  Widget _buildInfoText(String label, double amount, Color? color) {
     return Text(
       '$label: ${amount.toStringAsFixed(2)} L.E',
       style: TextStyle(fontSize: 18, fontWeight: FontWeight.bold, color: color),
     );
   }
 
-  Widget _buildBillTable(List<Bill> bills, Map<int, int> updatedPayments) {
+  Widget _buildBillTable(List<Bill> bills, Map<int, double> updatedPayments) {
     return DataTable(
       columnSpacing: 12.0,
       columns: [
@@ -187,7 +187,7 @@ class _PaymentDialogState extends State<PaymentDialog> {
         DataColumn(label: Text('المتبقي')),
       ],
       rows: bills.map((bill) {
-        int remainingAmount = bill.total_price - updatedPayments[bill.id]!;
+        double remainingAmount = bill.total_price - updatedPayments[bill.id]!;
         return DataRow(cells: [
           DataCell(Text(bill.id.toString())),
           DataCell(Text('${bill.total_price.toStringAsFixed(2)} L.E')),
